requirements.txt
#requirements.txt

waymo-open-dataset-tf-2-12-0

pandas
pyarrow
numpy
scikit-learn

#Visualization
matplotlib

#Image Handling
Pillow
